refactor: adjust structure for components;
