+++ src/components/index.js
@@ -1,0 +1,5 @@
+// Components barrel exports for easier imports
+export * from './pages';
+export * from './layout';
+export * from './tables';
+export * from './modals';
