+++ src/components/layout/index.js
@@ -1,0 +1,4 @@
+// Layout Components
+export { default as EditorContent } from './EditorContent';
+export { default as EditorHeader } from './EditorHeader';
+export { default as ErrorDialog } from './ErrorDialog';
