+++ src/components/modals/index.js
@@ -1,0 +1,5 @@
+// Modal Components
+export { default as ClanMemberModal } from './ClanMemberModal';
+export { default as ResourceModal } from './ResourceModal';
+export { default as RetainerModal } from './RetainerModal';
+export { default as SpouseModal } from './SpouseModal';
