+++ src/components/pages/index.js
@@ -1,0 +1,3 @@
+// Page Components
+export { default as EditorPage } from './EditorPage';
+export { default as FileUploadPage } from './FileUploadPage';
