+++ src/components/tables/ClanMemberTable.jsx
@@ -1,11 +1,11 @@
-import useUIStore from '../store/useUIStore';
-import useGameDataStore from '../store/useGameDataStore';
+import useUIStore from '../../store/useUIStore';
+import useGameDataStore from '../../store/useGameDataStore';
 
-function SpouseTable() {
-	const { openSpouseModal } = useUIStore();
-	const { getSpousesData } = useGameDataStore();
-	
-	const spousesData = getSpousesData();
+function ClanMemberTable() {
+	const { openClanMemberModal } = useUIStore();
+	const { getCurrentData } = useGameDataStore();
+
+	const data = getCurrentData('clanMembers');
 
 	return (
 		<div className="overflow-x-auto">
@@ -31,26 +31,26 @@ function SpouseTable() {
 					</tr>
 				</thead>
 				<tbody className="bg-white divide-y divide-gray-200">
-					{spousesData.map((spouse, index) => (
-						<tr key={spouse.name + index} className="hover:bg-gray-50">
-							<td className="px-3 py-4 whitespace-nowrap text-sm font-medium text-gray-900">{spouse.name}</td>
-							<td className="px-3 py-4 whitespace-nowrap text-sm text-gray-600">{spouse.age}</td>
-							<td className="px-3 py-4 whitespace-nowrap text-sm text-gray-600">{spouse.literature}</td>
-							<td className="px-3 py-4 whitespace-nowrap text-sm text-gray-600">{spouse.martial}</td>
-							<td className="px-3 py-4 whitespace-nowrap text-sm text-gray-600">{spouse.commerce}</td>
-							<td className="px-3 py-4 whitespace-nowrap text-sm text-gray-600">{spouse.art}</td>
-							<td className="px-3 py-4 whitespace-nowrap text-sm text-gray-600">{spouse.strategy}</td>
-							<td className="px-3 py-4 whitespace-nowrap text-sm text-gray-600">{spouse.reputation}</td>
-							<td className="px-3 py-4 whitespace-nowrap text-sm text-gray-600">{spouse.luck}</td>
-							<td className="px-3 py-4 whitespace-nowrap text-sm text-gray-600">{spouse.charm}</td>
-							<td className="px-3 py-4 whitespace-nowrap text-sm text-gray-600">{spouse.health}</td>
-							<td className="px-3 py-4 whitespace-nowrap text-sm text-gray-600">{spouse.talent}</td>
-							<td className="px-3 py-4 whitespace-nowrap text-sm text-gray-600">{spouse.talentValue}</td>
-							<td className={`px-3 py-4 whitespace-nowrap text-sm ${spouse.skill === 'None' ? 'text-gray-500' : 'text-gray-600'}`}>{spouse.skill}</td>
-							<td className="px-3 py-4 whitespace-nowrap text-sm text-gray-600">{spouse.skillValue}</td>
+					{data.map((character, index) => (
+						<tr key={character.name + index} className="hover:bg-gray-50">
+							<td className="px-3 py-4 whitespace-nowrap text-sm font-medium text-gray-900">{character.name}</td>
+							<td className="px-3 py-4 whitespace-nowrap text-sm text-gray-600">{character.age}</td>
+							<td className="px-3 py-4 whitespace-nowrap text-sm text-gray-600">{character.literature}</td>
+							<td className="px-3 py-4 whitespace-nowrap text-sm text-gray-600">{character.martial}</td>
+							<td className="px-3 py-4 whitespace-nowrap text-sm text-gray-600">{character.commerce}</td>
+							<td className="px-3 py-4 whitespace-nowrap text-sm text-gray-600">{character.art}</td>
+							<td className="px-3 py-4 whitespace-nowrap text-sm text-gray-600">{character.strategy}</td>
+							<td className="px-3 py-4 whitespace-nowrap text-sm text-gray-600">{character.reputation}</td>
+							<td className="px-3 py-4 whitespace-nowrap text-sm text-gray-600">{character.luck}</td>
+							<td className="px-3 py-4 whitespace-nowrap text-sm text-gray-600">{character.charm}</td>
+							<td className="px-3 py-4 whitespace-nowrap text-sm text-gray-600">{character.health}</td>
+							<td className="px-3 py-4 whitespace-nowrap text-sm text-gray-600">{character.talent}</td>
+							<td className="px-3 py-4 whitespace-nowrap text-sm text-gray-600">{character.talentValue}</td>
+							<td className={`px-3 py-4 whitespace-nowrap text-sm ${character.skill === 'None' ? 'text-gray-500' : 'text-gray-600'}`}>{character.skill}</td>
+							<td className="px-3 py-4 whitespace-nowrap text-sm text-gray-600">{character.skillValue}</td>
 							<td className="px-3 py-4 whitespace-nowrap text-sm font-medium">
 								<button 
-									onClick={() => openSpouseModal(spouse, index)} 
+									onClick={() => openClanMemberModal(character, index)} 
 									className="text-gray-800 hover:text-gray-900 underline hover:no-underline cursor-pointer"
 								>
 									Edit
@@ -64,4 +64,4 @@ function SpouseTable() {
 	);
 }
 
-export default SpouseTable;
+export default ClanMemberTable;
