+++ src/components/tables/index.js
@@ -1,0 +1,5 @@
+// Table Components
+export { default as ClanMemberTable } from './ClanMemberTable';
+export { default as ResourceTable } from './ResourceTable';
+export { default as RetainerTable } from './RetainerTable';
+export { default as SpouseTable } from './SpouseTable';
